backend/requirements.txt
fastapi==0.108.0
uvicorn==0.24.0
gunicorn==21.2.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
asyncpg==0.29.0
aiosqlite==0.19.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
fastapi-csrf-protect>=0.2.0

# ML & Data
numpy==1.26.2
pandas==2.1.3
scikit-learn==1.3.2
xgboost==2.0.3
lightgbm==4.6.0
tensorflow==2.18.0
joblib==1.3.2

# API & Web
httpx==0.25.1
requests==2.31.0
aiohttp==3.9.1
beautifulsoup4==4.12.2

# Caching & Background Tasks
redis[asyncio]>=5.0.0
celery==5.3.4

# Monitoring & Logging
python-json-logger==2.0.7
sentry-sdk==1.38.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# LLM & AI
openai>=1.3.0

# Email
jinja2>=3.1.0
anthropic>=0.7.0

# Utils
python-dateutil==2.8.2
pytz==2023.3

# Payment Processing
stripe==8.0.0
